# docs/swagger.yaml
basePath: /v1
definitions:
...
  dto.Communion:
    properties:
      approved_at:
        description: ApprovedAt is the time when the communion was approved (if applicable)
        type: string
      approved_by:
        allOf:
        - $ref: '#/definitions/dto.User'
        description: ApprovedBy contains the admin user information (populated when
          needed)
      approved_by_user_id:
        description: ApprovedByUserID is the ID of the admin who approved the communion
        type: string
      communion_date:
        description: CommunionDate is the requested communion date (must be in the
          past)
        type: string
      created_at:
        description: CreatedAt is the time when the communion record was created
        type: string
      deleted_at:
        description: DeletedAt is the time when the communion record was soft deleted
        type: string
      id:
        description: ID is the unique identifier of the communion request
        type: string
      requested_at:
        description: RequestedAt is the time when the communion was requested
        type: string
      status:
        allOf:
        - $ref: '#/definitions/dto.CommunionStatus'
        description: Status is the current status of the communion request
      updated_at:
        description: UpdatedAt is the time when the communion record was last updated
        type: string
      user:
        allOf:
        - $ref: '#/definitions/dto.User'
        description: User contains the user information (populated when needed)
      user_id:
        description: UserID is the ID of the user who made the communion request
        type: string
    type: object
  dto.CommunionListResponse:
    properties:
      communions:
        items:
          $ref: '#/definitions/dto.Communion'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  dto.CommunionStatus:
    enum:
    - pending
    - approved
    - rejected
    type: string
    x-enum-varnames:
    - CommunionStatusPending
    - CommunionStatusApproved
    - CommunionStatusRejected
...
  dto.CreateCommunionRequest:
    properties:
      communion_date:
        description: 'CommunionDate is the desired communion date (format: YYYY-MM-DD)
          - must be in the past'
        type: string
    type: object
...
  dto.UpdateCommunionStatusRequest:
    properties:
      status:
        description: Status is the new status for the communion request
        type: string
    type: object
...
  dto.User:
    properties:
      childrens_name:
        description: ChildrensName is a list of children's names
        items:
          type: string
        type: array
      created_at:
        description: |-
          CreatedAt is the time when the user is created.
          It is automatically set when the user is created.
        type: string
      deleted_at:
        description: DeletedAt is the time the user was deleted.
        type: string
      education:
        description: Education is the user's education background
        type: string
      id:
        description: |-
          ID is the unique identifier of the user.
          It is automatically generated when the user is created.
        type: string
      job_title:
        description: JobTitle is the user's job title
        type: string
      lastname:
        description: LastName is the last name of the user.
        type: string
      marriage_status:
        description: MarriageStatus is the user's marriage status
        type: string
      name:
        description: Name is the first name of the user.
        type: string
      partner_name:
        description: PartnerName is the name of the user's partner (spouse)
        type: string
      phone_number:
        description: PhoneNumber is the Ethiopian phone number of the user.
        type: string
      role:
        description: Role is the user's role (admin, manager, user)
        type: string
      status:
        description: Status is the user's account status (PENDING, ACTIVE, INACTIVE)
        type: string
      telegram_id:
        description: TelegramID is the optional Telegram ID of the user.
        type: string
      updated_at:
        description: UpdatedAt is the time the user was last updated.
        type: string
    type: object
...
  model.ErrorResponse:
    properties:
      code:
        description: Code is the error code. It is not status code
        type: integer
      description:
        description: Description is the error description.
        type: string
      field_error:
        description: FieldError is the error detail for each field, if available that
          is.
        items:
          $ref: '#/definitions/model.FieldError'
        type: array
      message:
        description: Message is the error message.
        type: string
      stack_trace:
        description: |-
          StackTrace is the stack trace of the error.
          It is only returned for debugging
        type: string
    type: object
  model.FieldError:
    properties:
      description:
        description: Description is the error description for this field.
        type: string
      name:
        description: Name is the name of the field that caused the error.
        type: string
    type: object
  model.MetaData:
    properties:
      extra:
        description: Extra contains other response specific data
      page:
        description: Page  is used to identify the starting point to return rows from
          a result set
        type: integer
      per_page:
        description: PerPage is the number of records to return after filtering
        type: integer
      sort:
        description: Sort is used to sort the result set in ascending or descending
          order
        type: string
      total:
        description: Total is the total number of data without pagination
        type: integer
    type: object
...
  model.SuccessResponse:
    properties:
      data:
        description: Data contains the response data
      meta_data:
        allOf:
        - $ref: '#/definitions/model.MetaData'
        description: MetaData contains pagination or other metadata
      ok:
        description: OK is always true for successful responses
        example: true
        type: boolean
    type: object
host: localhost:8000
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A comprehensive authentication API supporting Ethiopian phone numbers
    with JWT-based security
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Kesis Ethiopian Phone Authentication API
  version: "1.0"
paths:
...
  /communion:
    post:
      consumes:
      - application/json
      description: 'Create a new communion request for the authenticated user. The
        communion date must be in the past (format: YYYY-MM-DD). Users can create
        multiple communion requests.'
      parameters:
      - description: Communion request details
        in: body
        name: communion
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCommunionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created communion request
          schema:
            $ref: '#/definitions/dto.Communion'
        "400":
          description: Bad request - validation errors (invalid date format, future
            date, etc.)
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create communion request
      tags:
      - Communion
  /communion/{id}:
    delete:
      description: Soft delete a communion request by ID. Requires 'communions:delete'
        permission (admin role only).
      parameters:
      - description: Communion ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted the communion request
          schema:
            $ref: '#/definitions/model.SuccessResponse'
        "400":
          description: Bad request - invalid communion ID
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden - requires admin role
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Communion request not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete communion request
      tags:
      - Communion
    get:
      description: Retrieve communion request information by ID. Users can view their
        own requests, admins can view any communion request.
      parameters:
      - description: Communion ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved communion request
          schema:
            $ref: '#/definitions/dto.Communion'
        "400":
          description: Bad request - invalid communion ID
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden - can only access own requests unless admin
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Communion request not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get communion request by ID
      tags:
      - Communion
    patch:
      consumes:
      - application/json
      description: Update a communion request. Users can only update their own pending
        requests. Only the communion date can be updated.
      parameters:
      - description: Communion ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated communion request details
        in: body
        name: communion
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCommunionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated communion request
          schema:
            $ref: '#/definitions/dto.Communion'
        "400":
          description: Bad request - validation errors (invalid date, non-pending
            request, etc.)
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden - can only update own pending requests
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Communion request not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update communion request
      tags:
      - Communion
  /communion/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status of a communion request to approved or rejected.
        Requires 'communions:manage' permission (admin role only). Only pending requests
        can have their status updated.
      parameters:
      - description: Communion ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Status update details (approved/rejected)
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCommunionStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated communion status
          schema:
            $ref: '#/definitions/dto.Communion'
        "400":
          description: Bad request - validation errors (invalid status, non-pending
            request, etc.)
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden - insufficient permissions (requires admin role)
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Communion request not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update communion request status
      tags:
      - Communion
  /communion/all:
    get:
      description: Retrieve a paginated list of all communion requests. Requires 'communions:list'
        permission (admin role only).
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 20, max: 100)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved all communion requests
          schema:
            $ref: '#/definitions/dto.CommunionListResponse'
        "400":
          description: Bad request - invalid pagination parameters
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden - insufficient permissions (requires admin role)
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all communion requests
      tags:
      - Communion
  /communion/pending:
    get:
      description: Retrieve a paginated list of pending communion requests that require
        approval. Requires 'communions:manage' permission (admin role only).
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 20, max: 100)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved pending communion requests
          schema:
            $ref: '#/definitions/dto.CommunionListResponse'
        "400":
          description: Bad request - invalid pagination parameters
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden - insufficient permissions (requires admin role)
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get pending communion requests
      tags:
      - Communion
  /communion/user:
    get:
      description: Retrieve a paginated list of communion requests for the authenticated
        user.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 20, max: 100)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user communion requests
          schema:
            $ref: '#/definitions/dto.CommunionListResponse'
        "400":
          description: Bad request - invalid pagination parameters
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's communion requests
      tags:
      - Communion
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"